refactor: consolidate test utilities and remove setupFile overhead (#609)
